Conflicts:
	src/view/View.java

+++ src/model/BasicTower.java
@@ -4,7 +4,7 @@ public class BasicTower extends Tower {
 	private int cost =1;
 	private int attack = 2;
 	private int range =10;
-	private int attackSpeed = 1;
+	private double attackSpeed = 1;
 	public BasicTower() {
 		super.setCost(cost);
 		super.setAttack(attack);
